Migrate to usecase pattern
