+++ libs/horizon/src/main/java/com/instructure/horizon/data/repository/ModuleItemOnlineRepository.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2026 - present Instructure, Inc.
+ *
+ *     Licensed under the Apache License, Version 2.0 (the "License");
+ *     you may not use this file except in compliance with the License.
+ *     You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *     Unless required by applicable law or agreed to in writing, software
+ *     distributed under the License is distributed on an "AS IS" BASIS,
+ *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *     See the License for the specific language governing permissions and
+ *     limitations under the License.
+ */
+package com.instructure.horizon.data.repository
+
+import com.instructure.canvasapi2.apis.ModuleAPI
+import com.instructure.canvasapi2.builders.RestParams
+import com.instructure.canvasapi2.models.CanvasContext
+import com.instructure.canvasapi2.models.ModuleObject
+import javax.inject.Inject
+
+class ModuleItemOnlineRepository @Inject constructor(
+    private val moduleApi: ModuleAPI.ModuleInterface,
+) {
+
+    suspend fun getModuleItemsForCourse(courseId: Long): List<ModuleObject> {
+        val params = RestParams(isForceReadFromNetwork = true)
+        return moduleApi.getFirstPageModulesWithItems(
+            CanvasContext.Type.COURSE.apiString,
+            courseId,
+            params,
+            includes = listOf("estimated_durations"),
+        ).dataOrThrow
+    }
+}
