+++ libs/horizon/src/main/java/com/instructure/horizon/domain/usecase/GetModuleItemsUseCase.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright (C) 2026 - present Instructure, Inc.
+ *
+ *     Licensed under the Apache License, Version 2.0 (the "License");
+ *     you may not use this file except in compliance with the License.
+ *     You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *     Unless required by applicable law or agreed to in writing, software
+ *     distributed under the License is distributed on an "AS IS" BASIS,
+ *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *     See the License for the specific language governing permissions and
+ *     limitations under the License.
+ */
+package com.instructure.horizon.domain.usecase
+
+import com.instructure.canvasapi2.models.ModuleObject
+import com.instructure.horizon.data.repository.ModuleItemOfflineRepository
+import com.instructure.horizon.data.repository.ModuleItemOnlineRepository
+import com.instructure.horizon.offline.OfflineSyncUseCase
+import com.instructure.pandautils.utils.FeatureFlagProvider
+import com.instructure.pandautils.utils.NetworkStateProvider
+import javax.inject.Inject
+
+class GetModuleItemsUseCase @Inject constructor(
+    private val onlineRepository: ModuleItemOnlineRepository,
+    private val offlineRepository: ModuleItemOfflineRepository,
+    networkStateProvider: NetworkStateProvider,
+    featureFlagProvider: FeatureFlagProvider,
+) : OfflineSyncUseCase<Long, List<ModuleObject>>(
+    syncEnabled = true,
+    networkStateProvider = networkStateProvider,
+    featureFlagProvider = featureFlagProvider,
+) {
+
+    override suspend fun execute(params: Long): List<ModuleObject> {
+        return if (shouldFetchFromNetwork()) {
+            onlineRepository.getModuleItemsForCourse(params).also {
+                if (shouldSync()) offlineRepository.saveModuleItem(params, it)
+            }
+        } else {
+            offlineRepository.getModuleItemsForCourse(params)
+        }
+    }
+}
